prevent empty tar when single file opened

+++ s3_batch.py
@@ -94,7 +94,8 @@ def create_batch_archives_and_send_to_s3(s3_buckets):
                     logging.info(directory)
                     logging.info(files)
                     elasticsearch_docs += create_tar_archives(bucket, directory, files, temporary_directory)
-                    buckets_to_sync.append(bucket)
+                    if elasticsearch_docs:
+                        buckets_to_sync.append(bucket)
 
         # this step syncs buckets to s3. aws cli is used for convenience
         for bucket in buckets_to_sync:
@@ -124,14 +125,15 @@ def create_tar_archives(bucket, directory, files, temporary_directory):
         batch_id = uuid4()
 
         s3_object_key = create_s3_object_key(s3_file_prefix, batch_id)
-        tar_info = create_archive(bucket, s3_object_key, archive_files, temporary_directory)
         remaining_files_to_archive = remaining_files_to_archive[len(archive_files):]
         archive_files = filter_opened_files(archive_files)
 
-        elasticsearch_docs += get_batch_elasticsearch_docs(bucket, s3_object_key, tar_info)
+        if archive_files:
+            tar_info = create_archive(bucket, s3_object_key, archive_files, temporary_directory)
+            elasticsearch_docs += get_batch_elasticsearch_docs(bucket, s3_object_key, tar_info)
 
-        for archive_file in archive_files:
-            os.remove(archive_file)
+            for archive_file in archive_files:
+                os.remove(archive_file)
 
     return elasticsearch_docs
 
@@ -200,8 +202,10 @@ def get_opened_files():
     for pid in psutil.pids():
         try:
             yield (file[0] for file in psutil.Process(pid).open_files())
-        except psutil.AccessDenied:
-            print("Access denied")
+        except psutil.AccessDenied as e:
+            pass
+            # logging.warning("Access denied while getting process opened files")
+            # logging.warning(str(e))
 
 
 def get_batch_files(directory, remaining_files_to_archive):
